pyproject.toml
[project]
name = "tg-marketing-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.17.0",
    "apscheduler-di>=0.1.0",
    "email-validator>=2.2.0",
    "gspread>=6.1.4",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "tortoise-orm>=0.24.0",
]
